refactor : VoucherCreator가 BigDecimal로 생성하도록 하고 각 금액 필드를 전부 BigDecimal로 수정

+++ src/main/java/com/prgms/VoucherApp/domain/FixedAmountVoucher.java
@@ -10,9 +10,9 @@ public class FixedAmountVoucher implements Voucher {
     private final UUID voucherId;
     private final BigDecimal fixedAmount;
 
-    public FixedAmountVoucher(UUID voucherId, double fixedAmount) {
+    public FixedAmountVoucher(UUID voucherId, BigDecimal fixedAmount) {
         this.voucherId = voucherId;
-        this.fixedAmount = BigDecimal.valueOf(fixedAmount);
+        this.fixedAmount = fixedAmount;
     }
 
     @Override
