+++ src/test/java/com/prgms/VoucherApp/domain/VoucherTest.java
@@ -19,24 +19,24 @@ class VoucherTest {
     @ParameterizedTest
     @CsvSource(value = {"1000:9000", "2000:8000", "0:10000", "500:9500", "10500:0"}, delimiter = ':')
     @DisplayName("고정 값 할인")
-    void fixedAmountDiscount(Long fixedAmount, Long result) {
+    void fixedAmountDiscount(BigDecimal fixedAmount, BigDecimal result) {
         fixedAmountVoucher = new FixedAmountVoucher(UUID.randomUUID(), fixedAmount);
-        assertThat(fixedAmountVoucher.discount(BigDecimal.valueOf(10000L))).isEqualTo(BigDecimal.valueOf(result));
+        assertThat(fixedAmountVoucher.discount(BigDecimal.valueOf(10000L))).isEqualTo(result);
     }
 
     @ParameterizedTest
     @CsvSource(value = {"10:9000", "15:8500", "100:0", "0:10000", "50:5000"}, delimiter = ':')
     @DisplayName("비율 값 할인")
-    void percentDiscount(Long percent, Long result) {
+    void percentDiscount(BigDecimal percent, BigDecimal result) {
         percentDiscountVoucher = new PercentDiscountVoucher(UUID.randomUUID(), percent);
-        assertThat(percentDiscountVoucher.discount(BigDecimal.valueOf(10000L))).isEqualTo(BigDecimal.valueOf(result));
+        assertThat(percentDiscountVoucher.discount(BigDecimal.valueOf(10000L))).isEqualTo(result);
     }
 
     @Test
     @DisplayName("비율 값 할인 무한 소수 예외 테스트")
     void percentDiscountException() {
         try {
-            percentDiscountVoucher = new PercentDiscountVoucher(UUID.randomUUID(), 3.33);
+            percentDiscountVoucher = new PercentDiscountVoucher(UUID.randomUUID(), BigDecimal.valueOf(3.33));
         } catch (ArithmeticException e) {
             Assertions.assertThatExceptionOfType(ArithmeticException.class)
                     .isThrownBy(() -> percentDiscountVoucher.discount(BigDecimal.valueOf(1)));
